Fix warnings on both clang and gcc 4.8.5 for unit tests when built wi…

+++ cpc/src/u32Table.cpp
@@ -77,7 +77,7 @@ void u32TableFree (u32Table * self) {
 void u32TableShow (u32Table * self) {
   Long tableSize = 1LL << self->lgSize;
   printf ("\nu32Table (%d valid bits; %lld of %lld slots occupied)\n",
-	  self->validBits, self->numItems, tableSize);
+	  self->validBits, (long long int) self->numItems, (long long int) tableSize);
   //  U32 * arr = self->slots;
   //  Long i;
   //  for (i = 0; i < tableSize; i++) {
@@ -101,7 +101,7 @@ void u32TableClear (u32Table * self) { // clear the table without resizing it
 
 void printU32Array (U32 * array, Long arrayLength) {
   Long i = 0;
-  printf ("\nu32Array [%lld]\n", arrayLength);
+  printf ("\nu32Array [%lld]\n", (long long int) arrayLength);
   for (i = 0; i < arrayLength; i++) {
     printf ("%d:\t%8X\n", (int) i, array[i]);    
   }
